refactor: buttonBox function이름 수정 &<br>태그 삭제
