+++ src/app/_component/main_home/Card.tsx
