+++ src/app/_component/main_home/Guide.tsx
