+++ src/app/page.tsx
@@ -4,9 +4,9 @@ import Input_Chat from "./_component/input/Input_Chat";
 import Footer from "./_component/footer/footer";
 import Toggle from "./_component/toggle/toggle";
 import Navigation from "./_component/navigation/page";
-import Guide from "./_component/_main01/Guide";
-import Card from "./_component/_main01/Card";
-import ButtonBox from "./_component/_main01/ButtonBox";
+import Guide from "./_component/main_home/Guide";
+import Card from "./_component/main_home/Card";
+import ButtonBox from "./_component/main_home/ButtonBox";
 
 import styles from "./_component/modal/modal.module.css";
 
@@ -28,7 +28,8 @@ export default function Home() {
       <hr />
       <Guide
           onClick={() => {}}>
-            쇼터디 가입하고<br />나에게 딱맞는 스터디 찾아요
+            <p>쇼터디 가입하고</p>
+            <p>나에게 딱맞는 스터디 찾아요</p>
       </Guide>
       <hr />
       <Card />
